Highlight interactive guide walkthrough snippets
How Kody works and Google OAuth swap the markdown body for a transcript, so body fence tokens never reached the visible CopyCodeBlocks. Collect those snippets on the origin, attach keyed highlight tokens, and paint them in the walkthrough.
